move conftest and site related views to new 'sites' app
