+++ server/apps/resources/json/paper_manifest_handler.json
