+++ server/apps/sites/views/__init__.py
@@ -1,0 +1,3 @@
+from .epa_site_views import EpaSiteSearchView, EpaSiteView
+from .profile_views import EpaPermissionView, RcraProfileView, SitePermissionView, SyncProfileView
+from .site_views import SiteApi, SiteList, SiteManifest, SyncSiteManifest
