+++ server/apps/trak/urls.py
@@ -2,21 +2,12 @@ from django.urls import include, path
 from rest_framework import routers
 
 from apps.trak.views import (
-    EpaSiteView,
     FederalWasteCodes,
-    HandlerSearch,
     ManifestHandlerView,
     ManifestView,
     MtnList,
     PullManifest,
-    RcraProfileView,
     SignManifestView,
-    SiteApi,
-    SiteList,
-    SiteManifest,
-    SitePermissionView,
-    SyncProfile,
-    SyncSiteManifest,
     TransporterView,
 )
 
@@ -24,24 +15,15 @@ manifest_router = routers.SimpleRouter()
 manifest_router.register(r"manifest", ManifestView)
 
 urlpatterns = [
-    # Rcra Profile
-    path("profile/<str:user>/sync", SyncProfile.as_view()),
-    path("profile/<str:user>", RcraProfileView.as_view()),
-    path("permission/<int:pk>", SitePermissionView.as_view()),
     # Manifest
     path("", include(manifest_router.urls)),
     path("manifest/pull", PullManifest.as_view()),
     path("manifest/sign", SignManifestView.as_view()),
     path("mtn", MtnList.as_view()),
     path("mtn/<str:epa_id>", MtnList.as_view()),
-    # Site
-    path("site/", SiteList.as_view()),
-    path("site/<str:epa_id>", SiteApi.as_view()),
-    path("site/<str:epa_id>/manifest", SiteManifest.as_view()),
-    path("site/manifest/sync", SyncSiteManifest.as_view()),
-    # EpaSite
-    path("epa_site/search", HandlerSearch.as_view()),
-    path("epa_site/details/<int:pk>", EpaSiteView.as_view()),
+    # Handler
+    # path("epa_site/search", EpaSiteSearchView.as_view()),
+    # path("epa_site/details/<int:pk>", EpaSiteView.as_view()),
     path("transporter/<int:pk>", TransporterView.as_view()),
     path("mtnhandler/<int:pk>", ManifestHandlerView.as_view()),
     # code
