+++ server/apps/trak/views/__init__.py
@@ -1,5 +1,5 @@
-from .handler_view import EpaSiteView, HandlerSearch, ManifestHandlerView, TransporterView
+from ...sites.views.epa_site_views import EpaSiteSearchView, EpaSiteView
+from ...sites.views.profile_views import RcraProfileView, SitePermissionView, SyncProfileView
+from .handler_view import ManifestHandlerView, TransporterView
 from .lookup_views import FederalWasteCodes
 from .manifest_view import ManifestView, MtnList, PullManifest, SignManifestView
-from .rcra_profile_view import RcraProfileView, SitePermissionView, SyncProfile
-from .site_view import SiteApi, SiteList, SiteManifest, SyncSiteManifest
